* [test] use weex-wd pkg
