+++ package.json
@@ -82,7 +82,8 @@
     "weex-components": "^0.2.0",
     "weex-picker": "^0.1.0",
     "weex-rax-framework": "0.1.0",
-    "weex-vue-framework": "2.1.8-weex.1"
+    "weex-vue-framework": "2.1.8-weex.1",
+    "weex-wd": "^1.0.1"
   },
   "devDependencies": {
     "xml2map": "^1.0.2",
