--- test/scripts/test.js
@@ -1,1 +1,0 @@
-require("./weex.js")({scheme:"testpage"})
